requirements.txt
python-redis-lock==3.2.0
redis==2.10.6
